A scalable implementation of the Play Billing Library
The Android developer community lacked guidance on how to implement Play
Billing in a scalable
fashion. Present usage of the Play Billing Library are not scalable as
projects grow to include
multiple teams working on multiple features. Instead, these
implementations are built in such a way
that the Play BillingClient and its peripherals tend to have tentacles
everywhere in the rest of
the app, requiring almost everyone on an Android development team to
have some familiarity with
how Play Billing works. Additionally, implementations tend to be buggy
because too much is
happening in too many places within the app.

This sample solves this problem by presenting the community a scalable
architecture.
The sample works well for projects ranging from the very small to the
very large.

design doc: go/scalable-iab-sample
b/122146259

Change-Id: I6f562e1e77788a4ed6fe0fcceed050c68dd8a5b2